Working Upgrade system

+++ game/Weapon.cpp
@@ -1323,7 +1323,7 @@ void rvWeapon::Save ( idSaveGame *savefile ) const {
 	projectileEnt.Save ( savefile );
 	savefile->WriteVec3		( ejectOffset );	// cnicholson: Added unsaved var
 
-	savefile->WriteInt		( fireRate );
+	savefile->WriteFloat		( fireRate );
 	savefile->WriteFloat	( spread );
 	// savefile->WriteInt	( nextAttackTime ); // cnicholson: This is set to 0 in restore, so don't save it
 
@@ -1455,7 +1455,7 @@ void rvWeapon::Restore ( idRestoreGame *savefile ) {
 	projectileEnt.Restore ( savefile );
 	savefile->ReadVec3		( ejectOffset );	// cnicholson: Added unrestored var
 
-	savefile->ReadInt		( fireRate );
+	savefile->ReadFloat		( fireRate );
 	savefile->ReadFloat		( spread );
 	nextAttackTime = 0;
